ring/installer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ring-installer"
version = "0.1.0"
description = "Multi-platform installer for Ring AI skills library"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Lerian Studio", email = "contact@lerian.studio"}
]
maintainers = [
    {name = "Fred Amaral", email = "fred@fredamaral.com.br"}
]
keywords = [
    "ai",
    "skills",
    "installer",
    "claude",
    "cursor",
    "cline",
    "factory"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/lerianstudio/ring"
Documentation = "https://github.com/lerianstudio/ring#readme"
Repository = "https://github.com/lerianstudio/ring"
Issues = "https://github.com/lerianstudio/ring/issues"

[project.scripts]
ring-installer = "ring_installer.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ring_installer*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
